+++ cmd/local/jvmcollect/ttopcollect_test.go
@@ -77,6 +77,7 @@ func TestTtopCollects(t *testing.T) {
 	ttopArgs := jvmcollect.TtopArgs{
 		PID:      pid,
 		Interval: interval,
+		TempDir:  outDir,
 	}
 	if err := jvmcollect.OnLoop(ttopArgs, duration, outDir, ttopService, timeTicker); err != nil {
 		t.Fatalf("unable to collect %v", err)
@@ -132,10 +133,12 @@ func TestTtopExec(t *testing.T) {
 			t.Log("Process killed successfully.")
 		}
 	}()
+	tmp := t.TempDir()
 
 	ttopArgs := jvmcollect.TtopArgs{
 		PID:      cmd.Process.Pid,
 		Interval: 1,
+		TempDir:  tmp,
 	}
 	if err := ttop.StartTtop(ttopArgs); err != nil {
 		t.Error(err.Error())
@@ -153,6 +156,7 @@ func TestTtopExecHasNoPidToFind(t *testing.T) {
 	ttopArgs := jvmcollect.TtopArgs{
 		PID:      89899999999,
 		Interval: 1,
+		TempDir:  t.TempDir(),
 	}
 	if err := ttop.StartTtop(ttopArgs); err != nil {
 		t.Error("expected an error on ttop but none happened")
@@ -168,6 +172,7 @@ func TestTtopExecHasNoPid(t *testing.T) {
 	ttopArgs := jvmcollect.TtopArgs{
 		PID:      -2,
 		Interval: 1,
+		TempDir:  t.TempDir(),
 	}
 	resp := ttop.StartTtop(ttopArgs)
 	time.Sleep(time.Duration(500) * time.Millisecond)
@@ -202,6 +207,7 @@ func TestTtopHasAndInvalidInterval(t *testing.T) {
 	ttopArgs := jvmcollect.TtopArgs{
 		PID:      cmd.Process.Pid,
 		Interval: 0,
+		TempDir:  t.TempDir(),
 	}
 	if err := ttop.StartTtop(ttopArgs); err == nil {
 		t.Error("expected ttop start to fail with interval 0")